requirements.txt
streamlit
PyMuPDF
google-generativeai==0.8.5
numpy
sentence-transformers==2.2.2

